# backend/problem-solver-backend/openapi.yaml
openapi: 3.0.3
info:
  title: Problem Solver API
  version: 0.1.0
  description: >
    Turns a raw problem into a sharpened brief plus three proposed AI applications.
    Flow: call /api/solve, then /api/blueprint for one chosen idea, then
    /api/image for that blueprint's render.

    The three app endpoints are public - no session needed, and no CSRF header
    since there is nothing to ride on. They are rate limited by client IP
    instead: 20 requests per hour, counted separately per endpoint. Accounts
    (Flask-Security, session-cookie based) still exist and gate /api/me;
    register or login to get the session + XSRF-TOKEN cookies, and echo the
    cookie value back in the X-XSRF-Token header on state-changing account
    calls.

    Every request body is capped at 64 KB (413 past that), and the free-text
    fields are capped at 2000 characters (400 past that).

servers:
  - url: http://localhost:5000
    description: Local dev (Flask). Frontend proxies /api/* here.

tags:
  - name: auth
    description: Flask-Security session auth
  - name: app
    description: Problem-solving endpoints

paths:
  /api/accounts/register:
    post:
      tags: [auth]
      summary: Register a user (auto-logs in; CONFIRMABLE is off)
      description: CSRF-exempt (unauth endpoint). Sets session + XSRF-TOKEN cookies on success.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Credentials'
      responses:
        '200': { description: Registered and logged in }
        '400': { description: Validation error (envelope with field errors) }

  /api/accounts/login:
    get:
      tags: [auth]
      summary: Seed the CSRF cookie
      description: >
        Returns the login form payload and sets the XSRF-TOKEN cookie even when
        logged out. Call this before POST login if you need the CSRF token.
      responses:
        '200': { description: XSRF-TOKEN cookie seeded }
    post:
      tags: [auth]
      summary: Log in
      description: CSRF-exempt (unauth endpoint). Sets session + XSRF-TOKEN cookies on success.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Credentials'
      responses:
        '200': { description: Logged in }
        '400': { description: Bad credentials (generic response) }

  /api/accounts/logout:
    post:
      tags: [auth]
      summary: Log out
      security:
        - sessionCookie: []
      responses:
        '200': { description: Logged out }

  /api/me:
    get:
      tags: [app]
      summary: Current user
      security:
        - sessionCookie: []
      responses:
        '200':
          description: The logged-in user
          content:
            application/json:
              schema:
                type: object
                properties:
                  id: { type: integer, example: 1 }
                  email: { type: string, example: test@test.com }
        '401': { description: Not authenticated }

  /api/solve:
    post:
      tags: [app]
      summary: Define a problem and propose an AI app
      description: >
        Runs the define -> solve agent pipeline. Public - no session or CSRF
        header required. Rate limited to 20 requests per hour per client IP.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [problem]
              properties:
                problem:
                  type: string
                  maxLength: 2000
                  example: All my images is a mess
      responses:
        '200':
          description: Brief plus three AI-app ideas
          content:
            application/json:
              schema:
                type: object
                properties:
                  brief: { $ref: '#/components/schemas/Brief' }
                  ideas:
                    type: array
                    items: { $ref: '#/components/schemas/Idea' }
        '400':
          description: Missing problem, or problem longer than 2000 characters
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
        '413': { $ref: '#/components/responses/PayloadTooLarge' }
        '429': { $ref: '#/components/responses/RateLimited' }
        '502':
          description: Agent returned invalid (non-JSON) output
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }

  /api/blueprint:
    post:
      tags: [app]
      summary: Blueprint one chosen idea
      description: >
        Runs the blueprint agent over the brief plus one idea from /api/solve.
        Returns architecture, build estimate, risks, and an image prompt to feed
        /api/image. Public - no session or CSRF header required. Rate limited to
        20 requests per hour per client IP.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [brief, idea]
              properties:
                brief: { $ref: '#/components/schemas/Brief' }
                idea: { $ref: '#/components/schemas/Idea' }
      responses:
        '200':
          description: Blueprint for the chosen idea
          content:
            application/json:
              schema:
                type: object
                properties:
                  blueprint: { $ref: '#/components/schemas/Blueprint' }
        '400':
          description: Missing brief or idea
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
        '413': { $ref: '#/components/responses/PayloadTooLarge' }
        '429': { $ref: '#/components/responses/RateLimited' }
        '502':
          description: Agent returned invalid (non-JSON) output
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }

  /api/image:
    post:
      tags: [app]
      summary: Render a concept screenshot
      description: >
        Generates a product screenshot from an image prompt, normally the
        image_prompt returned by /api/blueprint. Public - no session or CSRF
        header required. Rate limited to 20 requests per hour per client IP.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [image_prompt]
              properties:
                image_prompt:
                  type: string
                  maxLength: 2000
                  example: A polished SaaS dashboard showing an image library grouped by event
      responses:
        '200':
          description: The rendered image
          content:
            application/json:
              schema:
                type: object
                properties:
                  image:
                    type: string
                    description: PNG as a base64 data URI.
                    example: data:image/png;base64,iVBORw0KGgo...
        '400':
          description: Missing image prompt, or prompt longer than 2000 characters
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
        '413': { $ref: '#/components/responses/PayloadTooLarge' }
        '429': { $ref: '#/components/responses/RateLimited' }
        '502':
          description: Image generation failed
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }

components:
  responses:
    RateLimited:
      description: >
        Rate limit exceeded (20/hour per IP). Served by Flask-Limiter's default
        handler, so the body is HTML, not the Error schema.
    PayloadTooLarge:
      description: >
        Request body over 64 KB (MAX_CONTENT_LENGTH). Served by Flask's default
        handler, so the body is HTML, not the Error schema.

  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: session
      description: Flask session cookie, set by register/login.
    csrfHeader:
      type: apiKey
      in: header
      name: X-XSRF-Token
      description: Echo of the XSRF-TOKEN cookie value.

  schemas:
    Credentials:
      type: object
      required: [email, password]
      properties:
        email: { type: string, example: test@test.com }
        password: { type: string, example: Password123! }

    Brief:
      type: object
      description: Output of the define agent.
      properties:
        core_problem: { type: string }
        persona: { type: string }
        inputs:
          type: array
          items: { type: string }
        success_criteria: { type: string }

    Idea:
      type: object
      description: One of the three app ideas from the solve agent.
      properties:
        app_concept: { type: string }
        ai_approach: { type: string }
        required_inputs:
          type: array
          items: { type: string }
        azure_services:
          type: array
          items: { type: string }
        first_step: { type: string }

    Blueprint:
      type: object
      description: Output of the blueprint agent for one idea.
      properties:
        architecture: { type: string }
        build_estimate:
          type: object
          properties:
            weeks: { type: string, example: 2-4 weeks }
            drivers: { type: string }
        risks:
          type: array
          items:
            type: object
            properties:
              risk: { type: string }
              check: { type: string }
        image_prompt: { type: string }

    Error:
      type: object
      properties:
        error: { type: string }